@@ -12,7 +12,7 @@
     <RootNamespace>SkinnyJson.Unit.Tests</RootNamespace>
     <AssemblyName>SkinnyJson.Unit.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
-    <LangVersion>8</LangVersion>
+    <LangVersion>9</LangVersion>
     <FileAlignment>512</FileAlignment>
     <Nullable>enable</Nullable>
   </PropertyGroup>
